instead of event driven communication, now it's polling based
there's a class that takes care of updating buses, uses flags and lots of dictionaries to store the information as efficiant and possible.
when a function sends data, it thinks that it sent data but actually the message_sender just put it into the pool and will send it only at the next message interval.
haven't added data lock mechnism yet.