finalizing changes before preFinal assignment
added documentation to everything.
fixed a bug that the bus messed up reconnecting to the server if he couldn't connect at the first time.
changed the icon.
added proper loading screens (actually load things in the background while the loading window is displayed)
and a bunch of other small fixes